firefox: vimium: remove .direnv cruft
Removes some .direnv cruft that I accidentally checked in.

While I'm embarrassed I checked it in without noticing, I'm not
embarrased enough to go through the whole history and excise it. I've
audited the contents of what was checked in and confirmed it contained
nothing compromising for me or anyone else.