vm-lamu: add open-vm-tools functionality
Add support for open-vm-tools (the open source equivalent of vmware's
guest tools) on host vm-lamu. This includes the following functionality

  - shared clipboard copy-paste
  - shared folders
  - time synchronization

Upgrade vm-lamu linux kernel from the end of life 5.19 version to 6.0.
This change was prompted by updating NixOS channels to nixos-22.11 on
vm-lamu, which triggered an end of life error for the prior kernel
version. In general, a later kernel than the default is required because
vm-lamu runs on aarch64, which in my mind is somewhat related to the
fact that it is a virtual machine, hence it's lumped into this commit.