pkgs: move nix conf to system level
Move nix and nixpkgs configuration out of home-manager and into the
NixOS system configuration layer, in the pkgs directory, with other
overlays.

Source the dmenu-rs package from a commit in the master branch of
nixpkgs now that it's been merged.

Add a "compatibility" overlay so that my NixOS nixpkgs overlays are also
accessible by the nix command line tools. I followed the instructions on
the NixOS wiki on this, as linked in code comments.